refactor: удалит хук useLoginForm, вернет логику валидации в компонент loginForm
