+++ project/src/components/login-form/login-form.tsx
@@ -1,14 +1,27 @@
+import { ChangeEvent, FormEvent, useState } from 'react';
 import cn from 'classnames';
 
 import Spinner from '../spinner/spinner';
 
-import { useAppSelector } from '../../hooks';
-import useLoginForm from '../../hooks/use-login-form';
+import { useAppSelector, useAppDispatch } from '../../hooks';
 import { selectloginStatus } from '../../store/user-slice/user-slice';
+import { loginAction } from '../../store/user-slice/user-slice';
 
 import { FetchStatus } from '../../utils/const';
 
 import styles from './login-form.module.css';
+
+const REG_EXP_EMAIL = /^\S+@[aA-zZ]{2,10}\.[aA-zZ]{2,3}$/;
+const REG_EXP_PASSWORD = /([0-9]{1}[aA-zZ]{1})|([aA-zZ]{1}[0-9]{1})/i;
+
+interface LoginField {
+  value: string;
+  regexp: RegExp;
+  error: boolean;
+  errorText: string;
+}
+
+type InitialState = { [key: string]: LoginField };
 
 const fields = {
   email: {
@@ -22,12 +35,48 @@ const fields = {
 };
 
 function LoginForm(): JSX.Element {
+  const dispatch = useAppDispatch();
   const loginStatus = useAppSelector(selectloginStatus);
-
-  const { formState, handleChange, handleSubmit } = useLoginForm();
+  const [formState, setFormState] = useState<InitialState>({
+    email: {
+      value: '',
+      regexp: REG_EXP_EMAIL,
+      error: false,
+      errorText: 'Email is not entered correctly',
+    },
+    password: {
+      value: '',
+      regexp: REG_EXP_PASSWORD,
+      error: false,
+      errorText: 'Enter at least 1 number and 1 letter',
+    },
+  });
 
   const isPending = loginStatus === FetchStatus.Pending;
   const isValid = Object.values(formState).some(({ error }) => error);
+
+  const handleSubmit = (evt: FormEvent) => {
+    if (evt) evt.preventDefault();
+    const authData = { email: formState.email.value, password: formState.password.value };
+
+    dispatch(loginAction(authData));
+  };
+
+  const handleChange = (evt: ChangeEvent<HTMLInputElement>) => {
+    const { value, name } = evt.target;
+
+    const regExp = formState[name].regexp;
+    const isValid = regExp.test(value);
+
+    setFormState((prevState) => ({
+      ...prevState,
+      [name]: {
+        ...prevState[name],
+        error: !isValid,
+        value,
+      },
+    }));
+  };
 
   return (
     <section className="login">
